requirements.txt
fastapi
uvicorn[standard]
openai
numpy
pydantic[email]
pydantic_settings
requests
python-dotenv
Pillow
google-search-results
pymongo
motor
google-auth
passlib
python-jose
